Refactored FilteredMetricsService implementation
