+++ src/screens/datasharing/components/BaseTekUploadView.tsx
@@ -8,7 +8,8 @@ import {covidshield} from 'services/BackendService/covidshield';
 import {xhrError} from 'shared/fetch';
 import AsyncStorage from '@react-native-community/async-storage';
 import {INITIAL_TEK_UPLOAD_COMPLETE, ContagiousDateInfo, ContagiousDateType} from 'shared/DataSharing';
-import {EventTypeMetric, FilteredMetricsService} from 'services/MetricsService/FilteredMetricsService';
+import {FilteredMetricsService} from 'services/MetricsService/FilteredMetricsService';
+import {EventTypeMetric} from 'services/MetricsService/MetricsFilter';
 
 import {BaseDataSharingView} from './BaseDataSharingView';
